requirements.txt
asgiref==3.6.0
autopep8==2.0.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.1
flask-pydantic-api==0.9.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-schema-pydantic==1.2.4
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.7
pytest==7.3.1
pytest-asyncio==0.21.0
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.5.0
Werkzeug==2.3.1
